Correcciones idioma

+++ src/database/seeds/TruncateSeeder.php
@@ -13,15 +13,21 @@ class TruncateSeeder extends Seeder {
      */
     public function run()
     {
-        \Solunes\Payments\App\OnlineBankDeposit::truncate();
-        \Solunes\Payments\App\OnlineBank::truncate();
+        if(config('payments.online_banks')){
+            \Solunes\Payments\App\OnlineBankDeposit::truncate();
+            \Solunes\Payments\App\OnlineBank::truncate();
+        }
         \Solunes\Payments\App\PaymentTransaction::truncate();
-        \Solunes\Payments\App\PaymentShipping::truncate();
+        if(config('payments.shipping')){
+            \Solunes\Payments\App\PaymentShipping::truncate();
+        }
         \Solunes\Payments\App\PaymentItem::truncate();
         \Solunes\Payments\App\Payment::truncate();
-        \Solunes\Payments\App\ScheduledTransactionPayment::truncate();
-        \Solunes\Payments\App\ScheduledTransactionItem::truncate();
-        \Solunes\Payments\App\ScheduledTransaction::truncate();
+        if(config('payments.scheduled_transactions')){
+            \Solunes\Payments\App\ScheduledTransactionPayment::truncate();
+            \Solunes\Payments\App\ScheduledTransactionItem::truncate();
+            \Solunes\Payments\App\ScheduledTransaction::truncate();
+        }
         \Solunes\Payments\App\PaymentMethod::truncate();
     }
 }
